Put each agent in a separate .py file
